Time: 11 ms (35.06%), Space: 23 MB (21.2%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -4,7 +4,7 @@ public:
         
         int left=1;
         int right=*max_element(piles.begin(),piles.end());
-        double hours;
+        long long hours;
         int ans;
         
         while(left<=right){
